Add e2e tests for POST /auth/register

+++ src/common/utils/test-utils.ts
@@ -3,6 +3,10 @@ import { connection } from 'mongoose';
 import { internet } from 'faker';
 import { User } from '../../database/models/user/user';
 import { Config } from '../config';
+import { Constants } from '../constants';
+import { INestApplication } from '@nestjs/common';
+import { ExceptionFilter } from '../filters/exception.filter';
+import { TestingModule } from '@nestjs/testing';
 
 export class TestUtils {
     public static async connectToDatabase(): Promise<void> {
@@ -10,6 +14,13 @@ export class TestUtils {
     }
 
     public static async dropDatabase(): Promise<void> {
+        if (
+            Config.APP.MODE !== Constants.APP_MODE.TEST ||
+            connection.name !== 'test'
+        ) {
+            throw new Error('You cannot use it in the provided environment');
+        }
+
         await connection.db.dropDatabase();
     }
 
@@ -19,5 +30,17 @@ export class TestUtils {
             username: internet.userName(),
             password: internet.password(),
         };
+    }
+
+    public static async createTestApplication(
+        module: TestingModule,
+    ): Promise<INestApplication> {
+        const app = module.createNestApplication();
+
+        app.useGlobalFilters(new ExceptionFilter());
+
+        await app.init();
+
+        return app;
     }
 }
